requirements.txt
django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.1
python-dateutil==2.8.2 